Extract SubsonicMediaController from SubsonicRESTController
Move stream, hls, download, getCoverArt, getAvatar, getLyrics, and
getCaptions endpoints from SubsonicRESTController into a new
SubsonicMediaController, following the established Phase 2 controller
extraction pattern.

After the move, 10 @Autowired fields and 5 private helpers in
SubsonicRESTController became unreachable from any remaining endpoint
and were removed in the same commit. SubsonicRESTController shrinks
from 367 to 163 lines.

The shared utilities (wrapRequest, mapId, createPlayerIfNecessary,
createResponse, error) and the @ExceptionHandler(APIException.class)
are duplicated into the new controller; consolidation is tracked
under #98.

APIException and ErrorCode remain in SubsonicRESTController as they
are referenced by Spring Security infrastructure outside this package.

Verified: 766 tests pass, WAR builds clean, all 7 endpoints respond
correctly via live curl including the negative streamRole authorization
path returning the proper JAXB error envelope.

Part of #34, fixes #110.
